refactor: 정정내역관련한 컴포넌트들 modification폴더로 묶음
