+++ app/modificationHistory/page.jsx
@@ -1,6 +1,5 @@
 "use client";
-
-import ModificationHistory from "../../components/modificationHistory/modificationHistory";
+import ModificationHistory from "../../components/modification/modificationHistory/modificationHistory";
 
 export default function ModificationHistoryPage() {
   return <ModificationHistory />;
