+++ components/modification/modificationHistory/modificationHistory.jsx
@@ -1,9 +1,9 @@
 "use client";
 import styles from "./modificationHistory.module.scss";
-import Icon from "../common/icon/icon";
-import Modal from "../common/modal/modal";
+import Icon from "../../common/icon/icon";
+import Modal from "../../common/modal/modal";
 import { useState } from "react";
-import ModificationList from "../modificationList/modificationList";  // 추가된 부분
+import ModificationList from "../modificationList/modificationList";
 
 const datas = [
   {
